development/code/phase3/palmer-living-reality/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
bcrypt==4.0.1
Pillow==10.0.0
python-dateutil==2.8.2
pytest==7.4.0
gunicorn==21.2.0
email-validator==2.1.0
